// src/components/FieldForm/interface.ts
export type StoreValue = any;
export interface Store {
  [name: string]: StoreValue;
}
// ...
export type RuleRender = (form: FormInstance) => RuleObject;
// ...
type AggregationRule = BaseRule & Partial<ValidatorRule>;

interface ArrayRule extends Omit<AggregationRule, 'type'> {
  type: 'array';
  defaultField?: RuleObject;
}

export type RuleObject = AggregationRule | ArrayRule;

export type Rule = RuleObject | RuleRender;

export interface FormInstance<Values = any> {
  getFieldValue: (name: string) => any;
  getFieldsValue(): Values;
  setFieldsValue: (value: any) => void;
  validateFields: any;
  setCallbacks: (callbacks: Callbacks) => void;
  setInitialValues: (values: Store, init: boolean) => void;
  submit: () => void;
  resetFields: (fields?: string[]) => void;
}

export type EventArgs = any[];

export interface FieldEntity {
  onStoreChange: () => void;
  validateRules: () => any;
  validateError: (error: any) => void
  props: {
    name?: string;
    rules?: Rule[];
    initialValue?: any;
  };
}

export interface InternalFieldData {
  value: StoreValue;
}

export interface FieldData extends Partial<Omit<InternalFieldData, 'name'>> {
  name: string;
}

export interface ValidateErrorEntity<Values = any> {
  values: Values;
  errorFields: { name: string; errors: string[] }[];
  outOfDate: boolean;
}

export interface Callbacks<Values = any> {
  onValuesChange?: (changedValues: any, values: Values) => void;
  onFieldsChange?: (changedFields: FieldData[], allFields: FieldData[]) => void;
  onFinish?: (values: Values) => void;
  onFinishFailed?: (errorInfo: ValidateErrorEntity<Values>) => void;
}
// ...
export type InternalFormInstance = Omit<FormInstance, 'validateFields'> & {
  validateFields: any;

  validateTrigger?: string | string[] | false;

  /**
   * Form component should register some content into store.
   * We pass the `HOOK_MARK` as key to avoid user call the function.
   */
  // getInternalHooks: (secret: string) => InternalHooks;
};